requirements.txt
gevent
jinja2
pandas
parallel-ssh
paramiko
python-telegram-bot
scp
